把登录模块和原来的模块进行了整合： 1.合并user类，对原来的User增加了成员和函数。 2.对loginPage类进行修改：去除身份证验证；忘记密码功能完全由覆写实现 3.todo：其余模块可能由于User类的修改导致构造函数需要修改-待完成；单向加密：加盐或者换方式
